+++ src/app/(pages)/chat/page.tsx
@@ -1,7 +1,7 @@
-import { getServerSession } from 'next-auth';
+// import { getServerSession } from 'next-auth';
 import Link from 'next/link';
 import ChatListPage from '@/app/(pages)/chat/_components/ChatListPage';
-import { nextAuthOptions } from '@/app/api/auth/[...nextauth]/route';
+// import { nextAuthOptions } from '@/app/api/auth/[...nextauth]/route';
 import { Plus } from '@/assets/images/icons';
 import Header from '@/components/layout/Header';
 import PageLayout from '@/components/layout/PageLayout';
@@ -20,51 +20,51 @@ interface FetchProps {
   userId: string;
 }
 
-const room1 = [
-  {
-    roomId: 1,
-    roomName: 'General Chat',
-    roomType: 'challenge',
-    lastMessage: 'Hello everyone!',
-    lastTimestamp: '2024-11-07T15:30:00',
-    hasNewMessages: true,
-  },
-  {
-    roomId: 2,
-    roomName: 'Project Alpha',
-    roomType: 'challenge',
-    lastMessage: 'Please review the document.',
-    lastTimestamp: '2024-11-06T09:15:00',
-    hasNewMessages: false,
-  },
-  {
-    roomId: 3,
-    roomName: 'Random',
-    roomType: 'challenge',
-    lastMessage: 'Check out this meme!',
-    lastTimestamp: '2024-11-05T12:45:00',
-    hasNewMessages: true,
-  },
-];
+// const room1 = [
+//   {
+//     roomId: 1,
+//     roomName: 'General Chat',
+//     roomType: 'challenge',
+//     lastMessage: 'Hello everyone!',
+//     lastTimestamp: '2024-11-07T15:30:00',
+//     hasNewMessages: true,
+//   },
+//   {
+//     roomId: 2,
+//     roomName: 'Project Alpha',
+//     roomType: 'challenge',
+//     lastMessage: 'Please review the document.',
+//     lastTimestamp: '2024-11-06T09:15:00',
+//     hasNewMessages: false,
+//   },
+//   {
+//     roomId: 3,
+//     roomName: 'Random',
+//     roomType: 'challenge',
+//     lastMessage: 'Check out this meme!',
+//     lastTimestamp: '2024-11-05T12:45:00',
+//     hasNewMessages: true,
+//   },
+// ];
 
-const room2 = [
-  {
-    roomId: 4,
-    roomName: 'Support',
-    roomType: 'private',
-    lastMessage: null,
-    lastTimestamp: null,
-    hasNewMessages: false,
-  },
-  {
-    roomId: 5,
-    roomName: 'Development Team',
-    roomType: 'private',
-    lastMessage: 'Code review at 3 PM.',
-    lastTimestamp: '2024-11-07T10:00:00',
-    hasNewMessages: true,
-  },
-];
+// const room2 = [
+//   {
+//     roomId: 4,
+//     roomName: 'Support',
+//     roomType: 'private',
+//     lastMessage: null,
+//     lastTimestamp: null,
+//     hasNewMessages: false,
+//   },
+//   {
+//     roomId: 5,
+//     roomName: 'Development Team',
+//     roomType: 'private',
+//     lastMessage: 'Code review at 3 PM.',
+//     lastTimestamp: '2024-11-07T10:00:00',
+//     hasNewMessages: true,
+//   },
+// ];
 
 const fetchGroupChatListData = async ({ userId }: FetchProps) => {
   const groupChatListData = await socketApi.get<ChatItem[]>('/chat/list/group', {
@@ -85,7 +85,7 @@ const fetchChallengeChatListData = async ({ userId }: FetchProps) => {
 };
 
 const Page = async () => {
-  const session = await getServerSession(nextAuthOptions);
+  // const session = await getServerSession(nextAuthOptions);
 
   const groupChatListData = await fetchGroupChatListData({ userId: 'user1' });
   const challengeChatListData = await fetchChallengeChatListData({ userId: 'user1' });
